feat(operations): replace `full_record` and `__unset__` with a `put` operation
Scope reduction, per kriszyp's review on #2347: one operation instead of a flag
plus a per-record directive.

    { "operation": "put", "database": "dev", "table": "dog",
      "records": [{ "id": 1, "dog_name": "Penny" }] }

The stored record becomes exactly that, so an attribute the request omits is
removed. Creates a missing record and replaces an existing one — the same
`Table.put` REST's `PUT /Table/id` performs, so the same audit type, replication
shape, and retained `__createdtime__`. `update`/`upsert` keep merging, which is the
v4-compatible behaviour existing clients depend on and the reason this is a
separate operation rather than semantics changed in place.

This covers what studio#1643 actually needs: its row editor holds the whole record
already. Removed with the flag and the directive:

- `full_record` as a client-facing field on update/upsert
- `__unset__`, its reserved key, its validation, the attribute-collector
  divergence in both `insertUpdateValidate.js` and `insert.ts`, and the
  `getRecordAttributes` name expansion
- `takeUnsetAttributes` and its latch

Three problems retire with them, all raised in review and none now reachable:
`__unset__` degrading a merge to a whole-record put under contention (kept as a
feature request, see below), the reserved key reinterpreting a pre-existing
`__unset__` attribute with no migration, and the directive's interaction with
dynamic attribute expansion.

The internal marker is an ARGUMENT, not a field: `putRecords` calls
`upsertRecords(obj, true)`. Reading it off the request — which the first draft of
this commit did — let a client send `full_record: true` with an `update` and get a
replace, which makes the operation name stop describing the write and bypasses the
attribute-scoped denial, since that is keyed on the operation. A test pins it;
restoring the request read fails that test and nothing else.

Kept from the previous rounds, unchanged in substance:

- The attribute-scoped denial, now keyed on `op === write.put.name`. A replace
  removes what the request omits, and `checkAttributePerms` only sees what it
  supplies.
- The target-database resolution fix and its fail-closed backstop, including the
  numeric `table: 0` guard.
- The bulk-load opt-out keyed on the operation rather than a caller-supplied
  `action`. The `__unset__` half of that fix goes away with the directive; the
  pre-existing half — a direct request skipping every attribute check by attaching
  `action` — stays fixed and stays tested.

Wired as a first-class operation: `OPERATIONS_ENUM.PUT`, the dispatch map, the
replication catchup switch, `requiredPermissions` with `[insert, update]` (what
`upsert` and REST `PUT` both require), the grantable `data` operation group, and
the `new_attributes`/`txn_time` result cleanup. Returns `put_hashes`; a put never
skips.

Tests: `integrationTests/database/put-operation.test.ts`, renamed from
full-record-write — 18 probes, all passing. Every guard mutation-checked: making
put ignore the marker fails 5, reading the marker off the request fails 1, dropping
the attribute-scoped denial fails 1. 143 tests across the security, terminology,
auto-fields and patch-merge suites pass, plus 575 in northwind, which exercises the
CSV and bulk-load paths.

`__unset__` is not abandoned — HarperFast/harper#2350 is being reworked from "this
is broken" into "add it properly", with concurrency-safe removal at the patch
layer rather than a read-then-replace in the bridge.

Refs: HarperFast/studio#1643
