+++ dataLayer/harperBridge/bridgeUtility/insertUpdateValidate.js
@@ -4,7 +4,6 @@ const hdbUtils = require('../../../utility/common_utils.ts');
 const log = require('../../../utility/logging/harper_logger.ts');
 const { getDatabases } = require('../../../resources/databases.ts');
 const { ClientError } = require('../../../utility/errors/hdbError.ts');
-const { UNSET_ATTRIBUTES } = require('../../../utility/hdbTerms.ts');
 
 module.exports = insertUpdateValidate;
 
@@ -74,15 +73,6 @@ function insertUpdateValidate(writeObject) {
 		dups.add(hdbUtils.autoCast(record[hash_attribute]));
 
 		for (let attr in record) {
-			// `__unset__` names attributes to remove; it is not one itself. Collecting it would make an
-			// open (non-schemaDefined) table register `__unset__` as a real attribute, since
-			// upsertRecords feeds this list straight to Table.addAttributes.
-			//
-			// Deliberately does NOT add the names it removes, unlike the async twin's list in
-			// dataLayer/insert.ts: that one is the bulk-load attribute-permission input, where a removal
-			// must be authorized, while this one creates table attributes — registering an attribute in
-			// order to delete it would be backwards. Keep both comments in step.
-			if (attr === UNSET_ATTRIBUTES) continue;
 			attributes[attr] = 1;
 		}
 	});
