+++ integrationTests/database/put-operation/schema.graphql
