+++ utility/hdbTerms.ts
@@ -250,6 +250,7 @@ export const OPERATIONS_ENUM = {
 	INSERT: 'insert',
 	UPDATE: 'update',
 	UPSERT: 'upsert',
+	PUT: 'put',
 	SEARCH_BY_CONDITIONS: 'search_by_conditions',
 	SEARCH_BY_HASH: 'search_by_hash',
 	SEARCH_BY_ID: 'search_by_id',
@@ -877,14 +878,6 @@ export const TIME_STAMP_NAMES_ENUM = {
 
 /** Timestamp values */
 export const TIME_STAMP_NAMES = [CREATED_TIME, UPDATED_TIME] as const;
-
-/**
- * Reserved record key naming attributes to REMOVE, for `update`/`upsert`: `__unset__: ["age"]`.
- * A write is otherwise a merge, so an attribute can only be dropped by replacing the whole record
- * (`full_record`); this removes named attributes while still merging everything else.
- * Reserved like the timestamp keys above — it is never stored, and never becomes a table attribute.
- */
-export const UNSET_ATTRIBUTES = '__unset__';
 
 /**
  * This value is used to help evaluate whether or not a permissions translation error is related to old permissions values or if it could be another code-related bug/error.
