+++ utility/operationPermissions.ts
@@ -69,6 +69,7 @@ export const OPERATION_PERMISSION_GROUPS = {
 		OPERATIONS_ENUM.INSERT,
 		OPERATIONS_ENUM.UPDATE,
 		OPERATIONS_ENUM.UPSERT,
+		OPERATIONS_ENUM.PUT,
 		OPERATIONS_ENUM.DELETE,
 		// Bulk load
 		OPERATIONS_ENUM.CSV_DATA_LOAD,
